+++ src/ext-test/java/org/opentripplanner/ext/fares/impl/AtlantaFareServiceTest.java
@@ -8,6 +8,7 @@ import static org.opentripplanner.ext.fares.impl.AtlantaFareService.XPRESS_AGENC
 import static org.opentripplanner.model.plan.TestItineraryBuilder.newItinerary;
 
 import java.util.Collection;
+import java.util.Currency;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -17,11 +18,12 @@ import org.opentripplanner._support.time.ZoneIds;
 import org.opentripplanner.ext.fares.model.FareRuleSet;
 import org.opentripplanner.framework.geometry.WgsCoordinate;
 import org.opentripplanner.framework.i18n.NonLocalizedString;
+import org.opentripplanner.model.fare.ItineraryFares;
 import org.opentripplanner.model.plan.Leg;
 import org.opentripplanner.model.plan.Place;
 import org.opentripplanner.model.plan.PlanTestConstants;
 import org.opentripplanner.routing.core.FareType;
-import org.opentripplanner.routing.core.ItineraryFares;
+import org.opentripplanner.transit.model.basic.Money;
 import org.opentripplanner.transit.model.basic.TransitMode;
 import org.opentripplanner.transit.model.framework.FeedScopedId;
 import org.opentripplanner.transit.model.network.Route;
@@ -32,6 +34,7 @@ public class AtlantaFareServiceTest implements PlanTestConstants {
 
   public static final float DEFAULT_TEST_RIDE_PRICE = 3.49f;
   public static final float DEFAULT_RIDE_PRICE_IN_CENTS = DEFAULT_TEST_RIDE_PRICE * 100;
+  public static final Currency USD = Currency.getInstance("USD");
   private static AtlantaFareService atlFareService;
 
   @BeforeAll
@@ -205,8 +208,18 @@ public class AtlantaFareServiceTest implements PlanTestConstants {
    */
   private static void calculateFare(List<Leg> rides, float expectedFareInCents) {
     ItineraryFares fare = new ItineraryFares();
-    atlFareService.populateFare(fare, null, FareType.electronicRegular, rides, null);
+    atlFareService.populateFare(fare, USD, FareType.electronicRegular, rides, null);
     assertEquals(expectedFareInCents, fare.getFare(FareType.electronicRegular).amount());
+
+    var fareProducts = fare
+      .getItineraryProducts()
+      .stream()
+      .filter(fp -> fp.id().getId().equals(FareType.electronicRegular.name()))
+      .toList();
+
+    assertEquals(1, fareProducts.size());
+    var fp = fareProducts.get(0);
+    assertEquals(Money.usDollars((int) expectedFareInCents), fp.price());
   }
 
   private static Leg getLeg(String agencyId, long startTimeMins) {