+++ src/ext/java/org/opentripplanner/ext/fares/model/RiderCategory.java
@@ -1,5 +1,0 @@
-package org.opentripplanner.ext.fares.model;
-
-import javax.annotation.Nullable;
-
-public record RiderCategory(String id, String name, @Nullable String url) {}