+++ src/main/java/org/opentripplanner/routing/core/FareComponent.java
@@ -1,6 +1,7 @@
 package org.opentripplanner.routing.core;
 
 import java.util.List;
+import org.opentripplanner.model.plan.Leg;
 import org.opentripplanner.transit.model.basic.Money;
 import org.opentripplanner.transit.model.framework.FeedScopedId;
 
@@ -8,10 +9,12 @@ import org.opentripplanner.transit.model.framework.FeedScopedId;
  * <p>
  * FareComponent is a sequence of routes for a particular fare.
  * </p>
+ * @deprecated Because it exists only for backwards compatibility, and you should use the Fares V2
+ * type, namely {@link org.opentripplanner.model.fare.FareProduct}.
  */
-public record FareComponent(
-  FeedScopedId fareId,
-  String name,
-  Money price,
-  List<FeedScopedId> routes
-) {}
+@Deprecated
+public record FareComponent(FeedScopedId fareId, Money price, List<Leg> legs) {
+  public List<FeedScopedId> routes() {
+    return legs.stream().map(l -> l.getRoute().getId()).toList();
+  }
+}