+++ src/test/java/org/opentripplanner/graph_builder/module/islandpruning/AdaptivePruningTest.java
@@ -73,10 +73,7 @@ public class AdaptivePruningTest {
       // Add street data from OSM
       File osmFile = new File(osmPath);
       OsmProvider osmProvider = new OsmProvider(osmFile, true);
-      OsmModule osmModule = OsmModule
-        .of(osmProvider, graph)
-        .withCustomNamer(new TestCustomNamer())
-        .build();
+      OsmModule osmModule = OsmModule.of(osmProvider, graph).withEdgeNamer(new TestNamer()).build();
 
       osmModule.buildGraph();